requirements.txt
uvicorn
falcon
falcon_json
falcon-jsonify
dnspython

gunicorn
